refactor(location-manager): split tour controllers
