+++ apps/location-manager/packages/server/src/features/locations/controllers/core/index.ts
@@ -4,4 +4,6 @@ export * from "./field-suggestions.controller";
 export * from "./pending-suggestions.controller";
 export * from "./hierarchy.controller";
 export * from "./tours.controller";
+export * from "./tour-import.controller";
+export * from "./tour-media.controller";
 export * from "./types.controller";
